+++ timecontrol.c
@@ -16,7 +16,7 @@
  * along with this program; if not, write to the Free Software
  * Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA  02111-1307  USA
  *
- * $Id: timecontrol.c,v 1.22 2006/06/22 22:50:29 Superfly_Jon Exp $
+ * $Id: timecontrol.c,v 1.23 2006/09/21 22:24:06 Superfly_Jon Exp $
  */
 
 #include "config.h"
@@ -51,7 +51,7 @@
 #if USE_TIMECONTROL
 
 timecontrol tc=
-	{ 0, 0, 0, 0, 0, 0,  0, 1.0, 0, 0, 0};
+	{ 0, 0, 0, 0, 0, 0,  0, 1, 0, 0, 0};
 
 /* Key values */
 static char szTCPLAIN[] = N_("plain")
@@ -402,7 +402,7 @@ static void showTimeControl( timecontrol *ptc, int level, int levels )
 		30-strlen(szTCMULT),
 		"Scale old time by:",
 		szTCMULT,
-		time2str(ptc->dMultiplier, 0)
+		time2str((int)ptc->dMultiplier, 0)
 		);
